fix: camelcase test (#46)
Closes https://github.com/MeltanoLabs/target-snowflake/issues/40

The merge_from_stage and copy_from_stage accept a schema argument that
it uses to build a query that selects from the raw json files in the
stage using the json keys. Since the staging data is raw, we needed to
pass in the raw unconformed schema instead of the conformed schema. The
conformed schema converted the camelcase `Id` column to lower case `id`,
so the query it built was trying to select a key in the json file that
didnt exist. The query has lowercase and the file has camelcase.