chore: uncomment SnowflakeTargetSchemaUpdates test (#45)
Closes https://github.com/MeltanoLabs/target-snowflake/issues/43

It turns out that this test had been fixed by other changes. I suspect
it was part of https://github.com/MeltanoLabs/target-snowflake/pull/39.
Previously I had implemented a way to cache table properties but I
suspect that the cache was not being invalidated when the schema updated
so it was attempting to recreate an existing column since it didnt find
it in the cache.