fix: handle reserved words (#51)
Closes https://github.com/MeltanoLabs/target-snowflake/issues/48

Sqlalchemy handles some of the quoting for reserved words but we also
build sql manually so those dont get handled. Also when doing column
diffing like to decided whether a column exists or not we need to use
uppercase column names because we created them as uppercase with quotes
so case matters.

I found a bug with the COPY logic when the stream has schema updates.
Logged it in https://github.com/MeltanoLabs/target-snowflake/issues/53
and draft PR in https://github.com/MeltanoLabs/target-snowflake/pull/52.
Once that is fixed we can add a schema change to the test records with
no key properties so we assert alters work with reserved words when no
key properties are being used.