fix: handle colons by not conforming names (#55)
Closes https://github.com/MeltanoLabs/target-snowflake/issues/54

The original bug was related to unquoted column names when selecting
json data from the internal stage files. By using the clean property
name thats been cleansed by sqlalchemy it was resolved. It then revealed
a second bug that was due to the fact that the table had conformed
column names but the merge statement did not.

The conform_name method was being called at times that we didnt want it
to be. The singer schema was being passed to the merge ddl method after
being conformed so it couldnt find the columns in the table that it just
created. I also think we should avoid conforming too much, in this case
the transferwise variant simply succeeds with quotes around the column
name, this PR makes this variant act the same way.