chore: Bump sdk 0.29.0 (#73)
- bumps sdk
- removes all code that was migrated to the SDK
- fixes the testing hackiness to use the default testing framework from
the docs now
- had to remove the class fixture for resource/connection in the tests
otherwise `ScopeMismatch: You tried to access the function scoped
fixture runner with a class scoped request object` was raised

Pending:
- the updates to assert key properties are present in the records are
causing 3 failed tests because the column names are being conformed in
key properties but not in the record so it cant find them and is
erroring. UPDATE: I fixed in here and created
https://github.com/meltano/sdk/issues/1819. SECOND UPDATE: I opened
https://github.com/MeltanoLabs/target-snowflake/issues/74 and overrode
the `_singer_validate_message` logic to do nothing for now. Once
https://github.com/MeltanoLabs/target-snowflake/issues/74 is implemented
we can run the validation if update mode is chosen.