fix: handle column order mismatches (#117)
Closes https://github.com/MeltanoLabs/target-snowflake/issues/116

I created a new test that creates a table then sends singer records that
are in a different order. The current implementation causes the test to
fail because
https://github.com/MeltanoLabs/target-snowflake/actions/runs/6160685807/job/16718219357
its trying to put a string in a boolean column. After implementing
Derek's suggested fix to explicitly define column names in the copy into
statement, the tests pass.

cc @visch